Reorganize examples, update README, update pyproject.toml dev deps for pipecat-ai
