+++ examples/restaurant_reservation.py
@@ -4,17 +4,15 @@
 # SPDX-License-Identifier: BSD 2-Clause License
 #
 
-"""A dynamic restaurant reservation flow example using Direct Functions.
+"""A dynamic restaurant reservation flow example for Pipecat Flows.
 
-This example demonstrates a restaurant reservation system using dynamic flows with
-direct functions where conversation paths are determined at runtime.
-Direct functions combine the function definition and handler in a single function.
+This example demonstrates a restaurant reservation system using dynamic flows where
+conversation paths are determined at runtime. The flow handles:
 
-The flow handles:
 1. Greeting and party size collection
-2. Time preference gathering with availability checking
-3. Alternative time suggestions when unavailable
-4. Reservation confirmation
+2. Time preference gathering
+3. Availability checking with mock reservation system
+4. Confirmation and booking
 
 Multi-LLM Support:
 Set LLM_PROVIDER environment variable to choose your LLM provider.
@@ -48,11 +46,9 @@ from pipecat.transports.base_transport import BaseTransport, TransportParams
 from pipecat.transports.daily.transport import DailyParams
 from pipecat.transports.websocket.fastapi import FastAPIWebsocketParams
 from pipecat.utils.text.markdown_text_filter import MarkdownTextFilter
-
-sys.path.append(str(Path(__file__).parent.parent))
 from utils import create_llm
 
-from pipecat_flows import FlowManager, FlowResult, NodeConfig
+from pipecat_flows import FlowArgs, FlowManager, FlowResult, FlowsFunctionSchema, NodeConfig
 
 load_dotenv(override=True)
 
@@ -120,54 +116,72 @@ class TimeResult(FlowResult):
 
 
 # Function handlers
-async def collect_party_size(
-    flow_manager: FlowManager, size: int
-) -> tuple[PartySizeResult, NodeConfig]:
-    """
-    Record the number of people in the party.
-
-    Args:
-        size (int): Number of people in the party. Must be between 1 and 12.
-    """
-    # Result: the recorded party size
+async def collect_party_size(args: FlowArgs) -> tuple[PartySizeResult, NodeConfig]:
+    """Process party size collection."""
+    size = args["size"]
     result = PartySizeResult(size=size, status="success")
-
-    # Next node: time selection
     next_node = create_time_selection_node()
-
     return result, next_node
 
 
-async def check_availability(
-    flow_manager: FlowManager, time: str, party_size: int
-) -> tuple[TimeResult, NodeConfig]:
-    """
-    Check availability for requested time.
+async def check_availability(args: FlowArgs) -> tuple[TimeResult, NodeConfig]:
+    """Check reservation availability and return result."""
+    time = args["time"]
+    party_size = args["party_size"]
 
-    Args:
-        time (str): Requested reservation time in "HH:MM AM/PM" format. Must be between 5 PM and 10 PM.
-        party_size (int): Number of people in the party.
-    """
     # Check availability with mock API
     is_available, alternative_times = await reservation_system.check_availability(party_size, time)
 
-    # Result: availability status and alternative times, if any
     result = TimeResult(
         status="success", time=time, available=is_available, alternative_times=alternative_times
     )
 
-    # Next node: confirmation or no availability
     if is_available:
+        logger.debug("Time is available, transitioning to confirmation node")
         next_node = create_confirmation_node()
     else:
-        next_node = create_no_availability_node(alternative_times)
+        logger.debug(f"Time not available, storing alternatives: {result['alternative_times']}")
+        next_node = create_no_availability_node(result["alternative_times"])
 
     return result, next_node
 
 
-async def end_conversation(flow_manager: FlowManager) -> tuple[None, NodeConfig]:
-    """End the conversation."""
+async def end_conversation(args: FlowArgs) -> tuple[None, NodeConfig]:
+    """Handle conversation end."""
     return None, create_end_node()
+
+
+# Create function schemas
+party_size_schema = FlowsFunctionSchema(
+    name="collect_party_size",
+    description="Record the number of people in the party",
+    properties={"size": {"type": "integer", "minimum": 1, "maximum": 12}},
+    required=["size"],
+    handler=collect_party_size,
+)
+
+availability_schema = FlowsFunctionSchema(
+    name="check_availability",
+    description="Check availability for requested time",
+    properties={
+        "time": {
+            "type": "string",
+            "pattern": "^([5-9]|10):00 PM$",  # Matches "5:00 PM" through "10:00 PM"
+            "description": "Reservation time (e.g., '6:00 PM')",
+        },
+        "party_size": {"type": "integer"},
+    },
+    required=["time", "party_size"],
+    handler=check_availability,
+)
+
+end_conversation_schema = FlowsFunctionSchema(
+    name="end_conversation",
+    description="End the conversation",
+    properties={},
+    required=[],
+    handler=end_conversation,
+)
 
 
 # Node configurations
@@ -187,7 +201,7 @@ def create_initial_node(wait_for_user: bool) -> NodeConfig:
                 "content": "Warmly greet the customer and ask how many people are in their party. This is your only job for now; if the customer asks for something else, politely remind them you can't do it.",
             }
         ],
-        "functions": [collect_party_size],
+        "functions": [party_size_schema],
         "respond_immediately": not wait_for_user,
     }
 
@@ -203,7 +217,7 @@ def create_time_selection_node() -> NodeConfig:
                 "content": "Ask what time they'd like to dine. Restaurant is open 5 PM to 10 PM.",
             }
         ],
-        "functions": [check_availability],
+        "functions": [availability_schema],
     }
 
 
@@ -217,7 +231,7 @@ def create_confirmation_node() -> NodeConfig:
                 "content": "Confirm the reservation details and ask if they need anything else.",
             }
         ],
-        "functions": [end_conversation],
+        "functions": [end_conversation_schema],
     }
 
 
@@ -236,7 +250,7 @@ def create_no_availability_node(alternative_times: list[str]) -> NodeConfig:
                 ),
             }
         ],
-        "functions": [check_availability, end_conversation],
+        "functions": [availability_schema, end_conversation_schema],
     }
 
 
@@ -250,15 +264,15 @@ def create_end_node() -> NodeConfig:
                 "content": "Thank them and end the conversation.",
             }
         ],
-        "functions": [],
         "post_actions": [{"type": "end_conversation"}],
     }
 
 
+# Main setup
 async def run_bot(
     transport: BaseTransport, runner_args: RunnerArguments, wait_for_user: bool = False
 ):
-    """Run the restaurant reservation bot with direct functions."""
+    """Run the restaurant reservation bot."""
     stt = DeepgramSTTService(api_key=os.getenv("DEEPGRAM_API_KEY"))
     tts = CartesiaTTSService(
         api_key=os.getenv("CARTESIA_API_KEY"),
@@ -323,7 +337,7 @@ if __name__ == "__main__":
     import sys
 
     # Parse our custom argument first
-    parser = argparse.ArgumentParser(description="Restaurant reservation bot with direct functions")
+    parser = argparse.ArgumentParser(description="Restaurant reservation bot")
     parser.add_argument(
         "--wait-for-user",
         action="store_true",
