+++ pyproject.toml
@@ -42,7 +42,7 @@ dev = [
     "setuptools~=78.1.1",
     "setuptools_scm~=8.3.1",
     "python-dotenv>=1.0.1,<2.0.0",
-    "pipecat-ai[webrtc,daily,silero,deepgram,openai,cartesia,runner,anthropic,google,aws]>=0.0.79",
+    "pipecat-ai[webrtc,websocket,daily,silero,deepgram,cartesia,openai,anthropic,google,aws,runner]>=0.0.85",
 ]
 
 docs = [
